% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-AlignmentPairs-class.R
\docType{class}
\name{AlignmentPairs-class}
\alias{AlignmentPairs-class}
\alias{AlignmentPairs}
\alias{AlignmentPairs,GRanges,GRanges-method}
\title{Representation of an alignment pair}
\usage{
AlignmentPairs(query, subject, ...)

\S4method{AlignmentPairs}{GRanges,GRanges}(query, subject, ...)
}
\arguments{
\item{query}{vectorized object representing query}

\item{subject}{vectorized object representing subject}

\item{...}{Arguments to pass to constructor}
}
\value{
AlignmentPairs object
}
\description{
Pairs subclass

Constructor for AlignmentPairs
}
\details{
The AlignmentPairs class extends the
\link[S4Vectors]{Pairs-class}
}
\section{Constructor}{


\code{AlignmentPairs(query, subject, ...)}: Constructs an AlignmentPairs
object by aligning the vectors \code{query} and \code{subject}.
}

\seealso{
\link[S4Vectors]{Pairs-class}
}
