% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-AlignmentPairsList-class.R
\docType{class}
\name{AlignmentPairsList-class}
\alias{AlignmentPairsList-class}
\alias{AlignmentPairsList}
\alias{AlignmentPairsList,list-method}
\title{List of AlignmentPairs instances}
\usage{
AlignmentPairsList(obj, ...)

\S4method{AlignmentPairsList}{list}(obj)
}
\arguments{
\item{obj}{list}

\item{...}{ellipsis}
}
\value{
AlignmentPairsList object
}
\description{
Subclass of S4Vectors SimpleList, where each entry is
an AlignmentPairs object
}
\section{Constructor}{


\code{AlignmentPairsList(obj, ...)}: Constructs an
AlignmentPairsList object from the input
}

