% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairsList-class.R
\name{as.data.frame,AlignmentPairsList-method}
\alias{as.data.frame,AlignmentPairsList-method}
\title{as.data.frame}
\usage{
\S4method{as.data.frame}{AlignmentPairsList}(x, ..., .id = "id")
}
\arguments{
\item{x}{AlignmentPairsList object}

\item{...}{additional parameters to lapply}

\item{.id}{name of id column added by dplyr::bind_rows}
}
\value{
data.frame
}
\description{
Convert AlignmentPairsList to data.frame.
}
