% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairsList-class.R
\name{autoplot.AlignmentPairsList}
\alias{autoplot.AlignmentPairsList}
\title{autoplot.AlignmentPairsList}
\usage{
\method{autoplot}{AlignmentPairsList}(object, aes, vars, ..., which = "point")
}
\arguments{
\item{object}{AlignmentPairsList}

\item{aes}{aes mapping}

\item{vars}{variable mapping to facet plots}

\item{...}{additional parameters to ggplot function}

\item{which}{which plot to make. 'grid' option makes a scatter
plot with marginal densities}
}
\description{
autoplot.AlignmentPairsList
}
