% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairs-class.R
\name{breadthOfCoverage}
\alias{breadthOfCoverage}
\title{breadthOfCoverage}
\usage{
breadthOfCoverage(x)
}
\arguments{
\item{x}{GRangesList object}
}
\value{
integer representing the sum of range widths
}
\description{
Calculate breadth of coverage by sequence id
}
