% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairs-class.R
\name{countSubjectsByCoverage}
\alias{countSubjectsByCoverage}
\title{countSubjectsByCoverage}
\usage{
countSubjectsByCoverage(
  x,
  min.coverage = seq(0.25, 1, 0.25),
  min.match = 0.9,
  nmax = 5
)
}
\arguments{
\item{x}{AlignmentPairs object}

\item{min.coverage}{coverage cutoffs to apply to transcripts}

\item{min.match}{filter out hits with fraction matching bases less
than min.match}

\item{nmax}{maximum number of subjects to count}
}
\description{
countSubjectsByCoverage
}
