% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{deletions}
\alias{deletions}
\title{Get deletions from an alignment.}
\usage{
deletions(x, ...)
}
\arguments{
\item{x}{Object representation of a sequence alignment}

\item{...}{Additional argument list}
}
\value{
numeric
}
\description{
Get deletions from an alignment.
}
