% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{divergence}
\alias{divergence}
\title{Get divergence from an alignment}
\usage{
divergence(x, ...)
}
\arguments{
\item{x}{Object representation of a sequence alignment}

\item{...}{Additional argument list}
}
\value{
numeric
}
\description{
Get divergence from an alignment.
}
