% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecovr-package.R
\docType{package}
\name{genecovr-package}
\alias{genecovr}
\alias{genecovr-package}
\title{genecovr: Gene body coverage analysis to evaluate genome assemblies}
\description{
Gene body coverage analysis to evaluate genome assemblies.
}
\details{
The genecovr package provides functionality for evaluating gene
body coverage and other metrics in genome assemblies. The central
measurement unit is a gene or transcript model.
}
\section{Functions}{


autoplot: plot GRanges objects with gene body coverage metrics
}

\author{
\strong{Maintainer}: Per Unneberg \email{per.unneberg@nbis.se} (\href{https://orcid.org/0000-0001-5735-3315}{ORCID})

}
\keyword{internal}
