% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-GRanges-class.R
\docType{methods}
\name{insertions}
\alias{insertions}
\alias{insertions,GRanges-method}
\alias{insertions,GRanges,ANY}
\title{Get insertions from an alignment}
\usage{
insertions(x, ...)

\S4method{insertions}{GRanges}(x, as.bases = FALSE, ...)
}
\arguments{
\item{x}{Object representation of a sequence alignment}

\item{...}{Additional argument list}

\item{as.bases}{Get insertions as base pairs}
}
\value{
numeric
}
\description{
Get insertions from an alignment.
}
