% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-AlignmentPairs-class.R
\name{matches}
\alias{matches}
\alias{mismatches}
\alias{repmatches}
\alias{matches,AlignmentPairs-method}
\alias{mismatches,AlignmentPairs-method}
\alias{repmatches,AlignmentPairs-method}
\title{matches}
\usage{
matches(x, ...)

mismatches(x, ...)

repmatches(x, ...)

\S4method{matches}{AlignmentPairs}(x)

\S4method{mismatches}{AlignmentPairs}(x)

\S4method{repmatches}{AlignmentPairs}(x)
}
\arguments{
\item{x}{Object representation of a sequence alignment}

\item{...}{Additional argument list}
}
\value{
numeric

numeric

numeric
}
\description{
Get matches from an alignment.

Get mismatches from an alignment.

Get matches that are parts of repeats from an alignment.
}
