% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairsList-class.R
\name{plot.AlignmentPairsList}
\alias{plot.AlignmentPairsList}
\title{plot}
\usage{
\method{plot}{AlignmentPairsList}(x, ...)
}
\arguments{
\item{x}{object to plot}

\item{...}{additional arguments for autoplot}
}
\description{
plot an AlignmentPairsList
}
