% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-psl.R
\name{psl methods}
\alias{psl methods}
\title{methods-psl}
\description{
Methods for dealing with psl data
}
\details{
Methods for dealing with the Pattern Space Layout (psl)
file format. See http://genome.ucsc.edu/FAQ/FAQformat.html for
more information.
}
