% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-AlignmentPairs-class.R
\name{query}
\alias{query}
\alias{query<-}
\alias{query,AlignmentPairs-method}
\alias{query<-,AlignmentPairs,GRanges-method}
\title{Retrieve query from object}
\usage{
query(x, ...)

query(x) <- value

\S4method{query}{AlignmentPairs}(x)

\S4method{query}{AlignmentPairs,GRanges}(x) <- value
}
\arguments{
\item{x}{object representing an alignment}

\item{...}{Additional arguments}

\item{value}{object representing a query}
}
\value{
query object

object representing an alignment with an updated query slot
}
\description{
For any object with a query slot retrieve the query
object.

For any object with a query slot set the
query.
}
