% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sequence-data.R
\name{readFastaIndex}
\alias{readFastaIndex}
\title{readFastaIndex}
\usage{
readFastaIndex(fai, ...)
}
\arguments{
\item{fai}{fasta index file}

\item{...}{additional parameters to pass to GenomeInfoDb::Seqinfo}
}
\value{
Seqinfo object
}
\description{
Read a fasta index file
}
\examples{
fai <- system.file("extdata", "polished.fai",
       package="genecovr")
sinfo <- readFastaIndex(fai)

}
