% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-psl.R
\name{readPsl}
\alias{readPsl}
\title{readPsl}
\usage{
readPsl(
  filename,
  seqinfo.query = NULL,
  seqinfo.sbjct = NULL,
  metadata = list(),
  ...
)
}
\arguments{
\item{filename}{input filename}

\item{seqinfo.query}{Seqinfo object for query (transcripts)}

\item{seqinfo.sbjct}{Seqinfo object for subject (reference)}

\item{metadata}{metadata for AlignmentPairs result}

\item{...}{additional parameters}
}
\value{
AlignmentPairs object
}
\description{
Read psl output
}
\details{

}
\examples{
fn <- system.file("extdata", "transcripts2polished.psl",
      package="genecovr")
ap <- readPsl(fn)

}
