% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{score}
\alias{score}
\title{Retrieve score from object}
\usage{
score(x, ...)
}
\arguments{
\item{x}{object representing an alignment}

\item{...}{Additional arguments}
}
\value{
numeric
}
\description{
For any object with a score slot retrieve the score
object.
}
