% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairs-class.R
\name{summarizeGeneBodyCoverage}
\alias{summarizeGeneBodyCoverage}
\title{summarizeGeneBodyCoverage}
\usage{
summarizeGeneBodyCoverage(x, min.coverage = seq(0, 1, 0.05), min.match = 0.9)
}
\arguments{
\item{x}{AlignmentPairs object}

\item{min.coverage}{coverage cutoffs to apply to transcripts}

\item{min.match}{filter out hits with fraction matching bases less
than min.match}
}
\description{
summarizeGeneBodyCoverage
}
