% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairs-class.R
\name{geneBodyCoverage}
\alias{geneBodyCoverage}
\title{geneBodyCoverage}
\usage{
geneBodyCoverage(x, min.match = 0.9)
}
\arguments{
\item{x}{AlignmentPairs object}

\item{min.match}{filter out hits with fraction matching bases less
than min.match}
}
\value{
reduced and filtered DataFrame object where each row is a
transcript. seqnames and seqlengths columns are collected from
the corresponding seqinfo object. breadthOfCoverage is the
total width of all reduced regions of the transcript; dividing
this number by the total transcript length gives the coverage
fraction. The revmap lists the numerical ids of the matching
contigs which can be retrieved from the AlignmentPairs
subject. The hitCoverage lists the width of each reduced hit,
and hitStart and hitEnd provide the transcript coordinates of
these hits.
}
\description{
Calculate geneBodyCoverage of query sequences
}
\details{
Given an AlignmentPairs object, calculate gene body
coverages of query sequences.
}
