% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AlignmentPairs-class.R
\name{geneBodyCoverage}
\alias{geneBodyCoverage}
\title{geneBodyCoverage}
\usage{
geneBodyCoverage(x, min.match = 0.9)
}
\arguments{
\item{x}{AlignmentPairs object}

\item{min.match}{filter out hits with fraction matching bases less
than min.match}
}
\value{
reduced and filtered DataFrame object where each row is a
transcript. seqnames and seqlengths columns are collected from
the corresponding seqinfo object. breadthOfCoverage is the
total width of all reduced regions of the transcript and
corresponds to how much of a transcript is covered. Dividing
breadthOfCoverage by the total transcript length gives the
coverage fraction. depthOfCoverage sums all ranges and is an
estimate of how many times each range is present. Dividing the
depthOfCoverage by breadthOfCoverage indicates the
multiplicity of the query. For queries that map multiple
subjects, a value close to 1 indicates the query has been
split in several subjects, whereas a higher value indicates
sequence duplication at the subject level.\preformatted{The revmap column maps the output ranges to the input ranges
as lists of numerical ids. These ids can be used to retrieve
the corresponding AlignmentPairs ranges providing a link to
the subjects. The hitCoverage lists the width of each reduced
hit, and hitStart and hitEnd provide the transcript
coordinates of these hits.
}
}
\description{
Calculate geneBodyCoverage of query sequences
}
\details{
Given an AlignmentPairs object, calculate gene body
coverages of query sequences.
}
